fixed fonts - Added labels on date fields - implimented countries and cities dropdown
