+++ client/src/assests/scss/app.scss
@@ -230,3 +230,9 @@ tr:nth-child(odd){
     display: inline-block;
     margin-right: 5px;
 }
+
+input{
+    color: #616060 !important;
+    font-size: .8rem !important;
+    font-family: 'Quicksand', sans-serif !important;
+}
